data_cleaning/requirements.txt
openai==0.28.1
python-dotenv==1.0.0
huggingface-hub==0.20.3
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
tqdm==4.66.1 